Align tests and examples naming convention with maven-surefire-plugin (used by default). Move all examples to the examples package.
