+++ src/test/java/com/aerospike/mapper/examples/model/InterestType.java
@@ -1,4 +1,4 @@
-package com.aerospike.mapper.example.model;
+package com.aerospike.mapper.examples.model;
 
 public enum InterestType {
 	INTEREST_ONLY,
