+++ src/test/java/com/aerospike/mapper/examples/model/Property.java
@@ -1,4 +1,4 @@
-package com.aerospike.mapper.example.model;
+package com.aerospike.mapper.examples.model;
 
 import java.util.ArrayList;
 import java.util.List;
